+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Trades/TradeHistoryVM.cs
@@ -42,6 +42,7 @@ namespace TD.WPF.ViewModels.Trades
             TakeProfitOrderDeleteCommand = new TakeProfitOrderDeleteCommand(this);
             TakeProfitOrderEditCommand = new TakeProfitOrderEditCommand(this, windowDialogService);
 
+            TradeImages = new ObservableCollection<TradeImage>();
             TradeImageCategoryValues = GetValues<TradeImageCategory>();
             TradeImageCreateCommand = new TradeImageCreateCommand(this, windowDialogService);
             TradeImageDeleteCommand = new TradeImageDeleteCommand(this);