+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/IOpenFileContext.cs
@@ -1,0 +1,8 @@
+﻿namespace TD.WPF.ViewModels
+{
+    public interface IOpenFileContext
+    {
+        string OpenFileDialogFilter { get; }
+        void SetFilePath(string filePath);
+    }
+}