+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Views/Trades/TradeImageCreateView.xaml.cs
@@ -41,6 +41,7 @@ namespace TD.WPF.Views.Trades
         {
             if (_preselectedTrade != null)
             {
+                _viewModel.Trade = _preselectedTrade;
                 _viewModel.TradeId = _preselectedTrade.Id;
             }
             OnSourceInitialized();